examples/customer-support-agent/requirements.txt
fastaiagent>=0.1.1
python-dotenv>=1.0.0
